refactor: move autocmd set up to autocmds.lua
